+++ app/src/main/java/com/pokemonify/pokemonify/fragments/MainFragment.java
@@ -69,7 +69,7 @@ public class MainFragment extends Fragment {
     }
 
     public void setMyPokemon() {
-        int temp = Utils.getMyPokemon(getActivity());
+        long temp = Utils.getMyPokemon(getActivity());
         if (temp == -1) {
             currentMyPokemon = new PokemonDto(0, "Snorlax", 50, "", "Mouse", "", "lazy,fat", 18, 3, 15);
         } else {
